2019_09 solved + intcode improvements

+++ problems/src/2019/02.scala
@@ -5,17 +5,17 @@ object problem02 extends baseProblem {
 
   import adventOfCode.utils.intcode._
 
-  override def solve1(input: Input): Int = {
+  override def solve1(input: Input): Long = {
     runPatchedCodes(parseMemory(input), noun = 12, verb = 2)
   }
 
-  override def solve2(input: Input): Int = {
+  override def solve2(input: Input): Long = {
     val codes = parseMemory(input)
     val resultToFind = 19690720
 
     val solutions = for {
-      noun <- (0 to 99).iterator
-      verb <- (0 to 99).iterator
+      noun <- (0L to 99L).iterator
+      verb <- (0L to 99L).iterator
       if runPatchedCodes(codes, noun, verb) == resultToFind
     } yield 100 * noun + verb
